+++ ledger/ledger.go
@@ -136,18 +136,4 @@ func CloseChannel(ctx context.Context, ledgerClient ledgerPb.ChannelsClient, sig
 	return nil
 }
 
-func Sign(key ic.PrivKey, channelMessage proto.Message) ([]byte, error) {
-	raw, err := proto.Marshal(channelMessage)
-	if err != nil {
-		return nil, err
-	}
-	return key.Sign(raw)
-}
 
-func Verify(key ic.PubKey, channelMessage proto.Message, sig []byte) (bool, error) {
-	raw, err := proto.Marshal(channelMessage)
-	if err != nil {
-		return false, err
-	}
-	return key.Verify(raw, sig)
-}
